Переместил фрагменты в отдельный пакет, добавил установку заголовка при изменении фрагментов, новые иконки приложения
