+++ app/src/main/res/layout/fragment_about.xml
@@ -2,6 +2,6 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.aleksandersh.weather.AboutFragment">
+             tools:context="com.aleksandersh.weather.fragment.AboutFragment">
 
 </FrameLayout>
